requirements.txt
streamlit
mariadb
streamlit-authenticator
pyyaml